% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialid-package.R
\docType{package}
\name{spatialid-package}
\alias{spatialid}
\alias{spatialid-package}
\title{spatialid: Create Persistent Identifiers for Longitudinal Spatial Data}
\description{
Provides tools for creating persistent identifiers for spatial units observed across time. The package supports configurable rules for identifying continuity, boundary changes, and spatial lineages in longitudinal polygon datasets.
}
\author{
\strong{Maintainer}: Emre Cebeci \email{cebeciemre1@gmail.com}

Authors:
\itemize{
  \item Emre Cebeci \email{cebeciemre1@gmail.com}
}

}
\keyword{internal}
