% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialid_example.R
\name{spatialid_example}
\alias{spatialid_example}
\title{Create an example longitudinal spatial dataset}
\usage{
spatialid_example()
}
\value{
An sf object containing synthetic polygons.
}
\description{
Generates a small synthetic spatial dataset for testing
spatialid functions. The example contains spatial units
observed across multiple years.
}
\examples{
example_data <- spatialid_example()
example_data
}
