% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialid_example.R
\name{spatialid_example}
\alias{spatialid_example}
\title{Create an example longitudinal spatial dataset}
\usage{
spatialid_example()
}
\value{
An sf object containing synthetic polygons.
}
\description{
Generates synthetic polygon data showing common spatial
continuity problems: stable units, boundary changes,
renames, and splits.
}
\examples{
data <- spatialid_example()
plot(data["name"])
}
